feature: 研发商店详情相关页面 issue #1896 - 新增发布管理tab - 新增版本管理页面 - 支持修改调试项目
