+++ src/frontend/devops-atomstore/src/components/manage/detail/atom-detail/show.vue
@@ -2,7 +2,7 @@
     <article>
         <section class="show-detail">
             <img v-if="detail.logoUrl" :src="detail.logoUrl" class="detail-img">
-            <ul :class="[{ 'overflow': !hasShowAll }, 'detail-items']" ref="detail">
+            <ul class="detail-items" ref="detail">
                 <li class="detail-item">
                     <span class="item-name">{{ detail.name }}</span>
                 </li>
@@ -54,37 +54,11 @@
                     />
                 </li>
             </ul>
-            <span class="summary-all" @click="hasShowAll = true" v-if="isOverDes && !hasShowAll"> {{ $t('展开全部') }} </span>
-        </section>
-
-        <section class="show-version">
-            <span class="version-label">{{ $t('store.版本列表') }}</span>
-            <bk-button theme="primary"
-                class="version-button"
-                :disabled="disableAddVersion"
-                @click="editAtom('upgradeAtom', versionList[0].atomId)"
-            > {{ $t('store.新增版本') }} </bk-button>
-            <bk-table :data="versionList" :outer-border="false" :header-border="false" :header-cell-style="{ background: '#fff' }">
-                <bk-table-column :label="$t('store.版本')" prop="version"></bk-table-column>
-                <bk-table-column :label="$t('store.状态')" prop="atomStatus" :formatter="statusFormatter"></bk-table-column>
-                <bk-table-column :label="$t('store.创建人')" prop="creator"></bk-table-column>
-                <bk-table-column :label="$t('store.创建时间')" prop="createTime"></bk-table-column>
-                <bk-table-column :label="$t('store.操作')" width="120" class-name="handler-btn">
-                    <template slot-scope="props">
-                        <section v-show="!index">
-                            <span class="update-btn" v-if="props.row.atomStatus === 'INIT'" @click="editAtom('shelfAtom', props.row.atomId)"> {{ $t('store.上架') }} </span>
-                            <span class="update-btn"
-                                v-if="progressStatus.indexOf(props.row.atomStatus) > -1" @click="routerProgress(props.row.atomId)"> {{ $t('store.进度') }} </span>
-                        </section>
-                    </template>
-                </bk-table-column>
-            </bk-table>
         </section>
     </article>
 </template>
 
 <script>
-    import { atomStatusMap } from '@/store/constants'
     import labelList from '../../../labelList'
 
     export default {
@@ -93,8 +67,7 @@
         },
 
         props: {
-            detail: Object,
-            versionList: Array
+            detail: Object
         },
 
         data () {
@@ -106,53 +79,7 @@
                 jobTypeMap: {
                     'AGENT': this.$t('store.编译环境'),
                     'AGENT_LESS': this.$t('store.无编译环境')
-                },
-                progressStatus: ['COMMITTING', 'BUILDING', 'BUILD_FAIL', 'TESTING', 'AUDITING'],
-                upgradeStatus: ['AUDIT_REJECT', 'RELEASED', 'GROUNDING_SUSPENSION'],
-                isOverDes: false,
-                hasShowAll: false
-            }
-        },
-
-        computed: {
-            atomStatusList () {
-                return atomStatusMap
-            },
-
-            disableAddVersion () {
-                const firstVersion = this.versionList[0] || {}
-                return this.upgradeStatus.indexOf(firstVersion.atomStatus) === -1
-            }
-        },
-
-        mounted () {
-            this.$nextTick(() => {
-                this.isOverDes = this.$refs.detail.scrollHeight > 290
-            })
-        },
-
-        methods: {
-            statusFormatter (row, column, cellValue, index) {
-                return this.$t(this.atomStatusList[cellValue])
-            },
-
-            routerProgress (id) {
-                this.$router.push({
-                    name: 'releaseProgress',
-                    params: {
-                        releaseType: 'upgrade',
-                        atomId: id
-                    }
-                })
-            },
-
-            editAtom (routerName, id) {
-                this.$router.push({
-                    name: routerName,
-                    params: {
-                        atomId: id
-                    }
-                })
+                }
             }
         }
     }
