+++ src/frontend/devops-atomstore/src/views/manage/detail/show.vue
@@ -1,21 +1,16 @@
 <template>
     <article class="manage-detail">
         <header class="manage-detail-header">
-            <router-link :to="showEdit ? { name: 'edit' } : ''"
-                :class="{ disable: !showEdit }"
-                :title="!showEdit && $t('store.只有处于审核驳回、已发布、上架中止和已下架的状态才允许修改基本信息')"
+            <router-link :to="detail.editFlag ? { name: 'edit' } : ''"
+                :class="{ disable: !detail.editFlag }"
+                :title="!detail.editFlag && $t('store.只有处于审核驳回、已发布、上架中止和已下架的状态才允许修改基本信息')"
             >{{ $t('store.编辑') }}</router-link>
         </header>
 
-        <main v-bkloading="{ isLoading }" class="detail-main">
+        <main class="detail-main">
             <component :is="`${$route.params.type}Show`"
-                v-if="!isLoading"
                 class="detail-show"
                 :detail="detail"
-                :version-list="versionList"
-                :pagination="pagination"
-                @pageChanged="pageChanged"
-                @pageLimitChanged="pageLimitChanged"
             ></component>
         </main>
     </article>
@@ -32,83 +27,10 @@
             imageShow
         },
 
-        data () {
-            return {
-                showEdit: false,
-                isLoading: true,
-                versionList: [],
-                pagination: {
-                    current: 1,
-                    count: 1,
-                    limit: 10
-                }
-            }
-        },
-
         computed: {
             ...mapGetters('store', {
                 'detail': 'getDetail'
             })
-        },
-
-        created () {
-            this.getVersionList()
-        },
-
-        methods: {
-            pageLimitChanged (currentLimit, prevLimit) {
-                if (currentLimit === this.pagination.limit) return
-
-                this.pagination.current = 1
-                this.pagination.limit = currentLimit
-                this.getVersionList()
-            },
-
-            pageChanged (page) {
-                if (page) this.pagination.current = page
-                this.getVersionList()
-            },
-
-            getVersionList () {
-                const methodMap = {
-                    atom: this.getAtomVersion,
-                    image: this.getImageVersion
-                }
-                const type = this.$route.params.type
-                const currentMethod = methodMap[type]
-                this.isLoading = true
-                currentMethod().catch((err) => {
-                    this.$bkMessage({ message: err.message || err, theme: 'error' })
-                }).finally(() => {
-                    this.isLoading = false
-                })
-            },
-
-            getAtomVersion () {
-                return this.$store.dispatch('store/requestVersionList', {
-                    atomCode: this.detail.atomCode
-                }).then((res) => {
-                    this.versionList = res.records || []
-                    const lastestVersion = this.versionList[0] || {}
-                    const lastestStatus = lastestVersion.atomStatus
-                    this.showEdit = ['AUDIT_REJECT', 'RELEASED', 'GROUNDING_SUSPENSION', 'UNDERCARRIAGED'].includes(lastestStatus)
-                })
-            },
-
-            getImageVersion () {
-                const postData = {
-                    imageCode: this.detail.imageCode,
-                    page: this.pagination.current,
-                    pageSize: this.pagination.limit
-                }
-                return this.$store.dispatch('store/requestImageVersionList', postData).then((res) => {
-                    this.versionList = res.records || []
-                    this.pagination.count = res.count
-                    const lastestVersion = this.versionList[0] || {}
-                    const lastestStatus = lastestVersion.imageStatus
-                    this.showEdit = ['AUDIT_REJECT', 'RELEASED', 'GROUNDING_SUSPENSION', 'UNDERCARRIAGED'].includes(lastestStatus)
-                })
-            }
         }
     }
 </script>
@@ -195,24 +117,6 @@
                     border-left: 2px solid #1592ff;
                     border-bottom: 2px solid #1592ff;
                 }
-            }
-        }
-        /deep/ .show-version {
-            margin-top: 40px;
-            .version-label {
-                color: #999;
-                display: flex;
-                align-items: center;
-                &::after {
-                    content: '';
-                    height: 1px;
-                    background: #ebedf0;
-                    margin-left: 16px;
-                    flex: 1;
-                }
-            }
-            .version-button {
-                margin: 25px 0;
             }
         }
     }
