+++ src/frontend/devops-atomstore/src/views/manage/release-manage/index.vue
@@ -1,0 +1,3 @@
+<template>
+    <router-view></router-view>
+</template>
