+++ src/frontend/devops-atomstore/src/views/manage/setting/member-setting.vue
@@ -8,7 +8,27 @@
         <section v-bkloading="{ isLoading }" class="g-scroll-table">
             <bk-table :data="memberList" :outer-border="false" :header-border="false" :header-cell-style="{ background: '#fff' }" v-if="!isLoading">
                 <bk-table-column :label="$t('store.成员')" prop="userName"></bk-table-column>
-                <bk-table-column :label="$t('store.调试项目')" prop="projectName"></bk-table-column>
+                <bk-table-column :label="$t('store.调试项目')">
+                    <template slot-scope="props">
+                        <section class="member-project">
+                            <template v-if="props.row.editing">
+                                <bk-select :disabled="false" v-model="props.row.projectCode" :loading="isLoadingProject" searchable style="width: 250px">
+                                    <bk-option v-for="option in projectList"
+                                        :key="option.projectCode"
+                                        :id="option.projectCode"
+                                        :name="option.projectName">
+                                    </bk-option>
+                                </bk-select>
+                                <i class="bk-icon icon-check-1" @click="saveChangeProject(props.row)"></i>
+                                <i class="bk-icon icon-close" @click="props.row.editing = false"></i>
+                            </template>
+                            <template v-else>
+                                <span>{{ props.row.projectName }}</span>
+                                <i class="bk-icon icon-edit2" @click="startEditProject(props.row)"></i>
+                            </template>
+                        </section>
+                    </template>
+                </bk-table-column>
                 <bk-table-column :label="$t('store.角色')" width="160" prop="type" :formatter="typeFormatter"></bk-table-column>
                 <bk-table-column :label="$t('store.描述')" prop="type" :formatter="desFormatter"></bk-table-column>
                 <bk-table-column :label="$t('store.操作')" width="120" class-name="handler-btn">
@@ -53,6 +73,7 @@
 <script>
     import { mapGetters } from 'vuex'
     import labelList from '@/components/labelList.vue'
+    import api from '@/api'
 
     export default {
         components: {
@@ -96,7 +117,8 @@
                     loading: false,
                     user: '',
                     id: ''
-                }
+                },
+                projectList: []
             }
         },
 
@@ -104,7 +126,26 @@
             ...mapGetters('store', {
                 'detail': 'getDetail',
                 'userInfo': 'getUserInfo'
-            })
+            }),
+
+            storeType () {
+                const typeMap = {
+                    atom: 'ATOM',
+                    image: 'IMAGE'
+                }
+                const type = this.$route.params.type
+                return typeMap[type]
+            },
+
+            storeCode () {
+                const keyMap = {
+                    atom: 'atomCode',
+                    image: 'imageCode'
+                }
+                const type = this.$route.params.type
+                const key = keyMap[type]
+                return this.detail[key]
+            }
         },
 
         created () {
@@ -112,6 +153,30 @@
         },
 
         methods: {
+            saveChangeProject (row) {
+                this.isLoading = true
+                const data = {
+                    projectCode: row.projectCode,
+                    storeCode: this.storeCode,
+                    storeType: this.storeType
+                }
+                api.requestChangeProject(data).then(() => {
+                    return this.initData()
+                }).catch(err => this.$bkMessage({ message: err.message || err, theme: 'error' })).finally(() => {
+                    this.isLoading = false
+                })
+            },
+
+            startEditProject (row) {
+                row.editing = true
+                row.isLoadingProject = true
+                this.$store.dispatch('store/requestProjectList').then((res) => {
+                    this.projectList = res || []
+                }).catch(err => this.$bkMessage({ message: err.message || err, theme: 'error' })).finally(() => {
+                    row.isLoadingProject = false
+                })
+            },
+
             requireRule (name) {
                 return {
                     required: true,
@@ -134,16 +199,13 @@
             saveMember () {
                 this.$refs.addForm.validate().then(() => {
                     this.isSaving = true
-                    const params = {
+                    const postData = {
                         type: this.addMemberObj.form.type,
-                        member: [this.addMemberObj.form.memberName]
+                        member: [this.addMemberObj.form.memberName],
+                        storeCode: this.storeCode,
+                        storeType: this.storeType
                     }
-                    const saveMethodMap = {
-                        atom: () => this.$store.dispatch('store/addAtomMember', { params: Object.assign(params, { storeCode: this.detail.atomCode }) }),
-                        image: () => this.$store.dispatch('store/requestAddImageMem', Object.assign(params, { storeCode: this.detail.imageCode }))
-                    }
-                    const type = this.$route.params.type
-                    saveMethodMap[type]().then(() => {
+                    api.requestAddMember(postData).then(() => {
                         this.closeAddMember()
                         this.initData()
                     }).catch(err => this.$bkMessage({ message: err.message || err, theme: 'error' })).finally(() => {
@@ -173,14 +235,17 @@
 
             initData () {
                 this.isLoading = true
-                const methodMap = {
-                    atom: () => this.$store.dispatch('store/requestMemberList', { atomCode: this.detail.atomCode }),
-                    image: () => this.$store.dispatch('store/requestImageMemList', this.detail.imageCode)
+                const data = {
+                    storeCode: this.storeCode,
+                    storeType: this.storeType
                 }
-                const type = this.$route.params.type
-                methodMap[type]().then((res) => {
+                api.getMemberList(data).then((res) => {
                     this.memberCount = res.length
-                    this.memberList = res || []
+                    this.memberList = (res || []).map(x => {
+                        x.editing = false
+                        x.isLoadingProject = false
+                        return x
+                    })
                 }).catch(err => this.$bkMessage({ message: err.message || err, theme: 'error' })).finally(() => {
                     this.isLoading = false
                 })
@@ -194,17 +259,15 @@
             },
 
             requestDeleteMember () {
-                const id = this.deleteObj.id
-                const methodMap = {
-                    atom: () => this.$store.dispatch('store/requestDeleteMember', { atomCode: this.detail.atomCode, id }),
-                    image: () => this.$store.dispatch('store/requestDeleteImageMem', { imageCode: this.detail.imageCode, id })
-                }
-
-                const type = this.$route.params.type
                 this.deleteObj.loading = true
-                methodMap[type]().then(() => {
+                const data = {
+                    id: this.deleteObj.id,
+                    storeCode: this.storeCode,
+                    storeType: this.storeType
+                }
+                api.requestDeleteMember(data).then(() => {
                     this.memberCount--
-                    const index = this.memberList.findIndex(x => x.id === id)
+                    const index = this.memberList.findIndex(x => x.id === this.deleteObj.id)
                     this.memberList.splice(index, 1)
                 }).catch(err => this.$bkMessage({ message: err.message || err, theme: 'error' })).finally(() => {
                     this.deleteObj.loading = false
@@ -232,6 +295,19 @@
             padding: 32px;
             /deep/ .bk-form-radio:not(:last-child) {
                 margin-right: 32px;
+            }
+        }
+        .member-project {
+            display: flex;
+            align-items: center;
+            .bk-icon {
+                font-size: 24px;
+                cursor: pointer;
+            }
+        }
+        .g-scroll-table {
+            /deep/ .bk-table .bk-table-body-wrapper {
+                height: calc(100% - 43px);
             }
         }
     }
